+++ src/pyzet/sample_config.py
@@ -11,14 +11,12 @@ SAMPLE_CONFIG_UNIX = f"""\
 {_header}
 repo: ~/zet
 editor: {C.vim_unix_path}
-git: {C.git_unix_path}
 """
 
 SAMPLE_CONFIG_WINDOWS = f"""\
 {_header}
 repo: ~/zet
 editor: {C.vim_win_path}
-git: {C.git_win_path}
 """