+++ tests/sample_config_test.py
@@ -24,7 +24,6 @@ def test_sample_config_unix(capsys):
 # to use program name directly if it's on your PATH.
 repo: ~/zet
 editor: /usr/bin/vim
-git: /usr/bin/git
 """
     )
     assert err == ''
@@ -49,7 +48,6 @@ def test_sample_config_windows(capsys):
 # to use program name directly if it's on your PATH.
 repo: ~/zet
 editor: C:/Program Files/Git/usr/bin/vim.exe
-git: C:/Program Files/Git/cmd/git.exe
 """
     )
     assert err == ''